Process XML DocBook with xsl stylesheets
docbook2(html|pdf) are designed for SGML DocBook and use parser which
are only capable of processing sgml files. Many new versions of these
tools also stop with an error when they start to process these files.

xmlto helps us to execute a XSLT processor with correct paths set to the
xsl stylesheets for docbook. It needs extra tools like fop to generate
more complex output files like pdf or ps, but fop doesn't support eps or
pdf as when generating pdf files. It is possible to use ps output and
conversation to pdf as workaround.