Try to track changes on images using a stamp
It is not easy to track changes on files in subdirectories using a
Makefile. A workaround is to use a stamp file which is `touch`ed when
changes to images were made. We still need to call make in the images
subdirectory to generate new images. Because it is not possible to
depend on the image phony target in a non phony target without
recreating the non phony target in every call of make, we must deoend
on it in a earlier (phony) target.
This can lead to errors in rare circumstances like when somebody
deletes a target file in images and then tries to create batman.pdf
using `make batman.pdf` instead of `make batman_doc`.